Extract methods for manipulating image to Utils module
